Group pagination and continuation token tests
